rename testTrees.ts to testTreeScenarios.ts
