+++ packages/dds/tree/src/test/feature-libraries/optional-field/optionalFieldSnapshots.test.ts
@@ -16,7 +16,7 @@ import {
 import { brand } from "../../../util/index.js";
 import { takeJsonSnapshot, useSnapshotDirectory } from "../../snapshots/index.js";
 // eslint-disable-next-line import/no-internal-modules
-import { createSnapshotCompressor } from "../../snapshots/testTrees.js";
+import { createSnapshotCompressor } from "../../snapshots/testTreeScenarios.js";
 import { TestNodeId } from "../../testNodeId.js";
 import { Change } from "./optionalFieldUtils.js";
 import { TestChange } from "../../testChange.js";
