+++ packages/dds/tree/src/test/feature-libraries/sequence-field/sequenceFieldSnapshots.test.ts
@@ -9,7 +9,7 @@ import { RevisionTagCodec } from "../../../core/index.js";
 import { SequenceField } from "../../../feature-libraries/index.js";
 import { takeJsonSnapshot, useSnapshotDirectory } from "../../snapshots/index.js";
 // eslint-disable-next-line import/no-internal-modules
-import { createSnapshotCompressor } from "../../snapshots/testTrees.js";
+import { createSnapshotCompressor } from "../../snapshots/testTreeScenarios.js";
 import { TestNodeId } from "../../testNodeId.js";
 import { testIdCompressor } from "../../utils.js";
 import { generatePopulatedMarks } from "./populatedMarks.js";
