+++ packages/dds/tree/src/test/snapshots/summary.spec.ts
@@ -6,7 +6,7 @@
 import { TreeCompressionStrategy } from "../../feature-libraries/index.js";
 import { SharedTreeFormatVersion, type SharedTreeOptions } from "../../shared-tree/index.js";
 import { useSnapshotDirectory } from "./snapshotTools.js";
-import { generateTestTrees } from "./testTrees.js";
+import { generateTestTrees } from "./testTreeScenarios.js";
 import { takeSummarySnapshot } from "./utils.js";
 
 describe("snapshot tests", () => {
