+++ packages/dds/tree/src/test/tree/detachedFieldIndex.spec.ts
@@ -24,7 +24,7 @@ import {
 } from "../../util/index.js";
 import { takeJsonSnapshot, useSnapshotDirectory } from "../snapshots/index.js";
 // eslint-disable-next-line import/no-internal-modules
-import { createSnapshotCompressor } from "../snapshots/testTrees.js";
+import { createSnapshotCompressor } from "../snapshots/testTreeScenarios.js";
 import { testIdCompressor, testRevisionTagCodec } from "../utils.js";
 
 const mintedTag = testIdCompressor.generateCompressedId();
